+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - ```bin/run_all_spiders.sh``` now discovers and runs spiders
 using ```spiders.py``` and ```spiderhost.py``` respectively
+- ```bin/run_all_spiders.sh``` can now run spiders directly
+or inside a docker container
 
 ### Removed